pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tap-excel"
version = "0.1"
dependencies = [
    "singer_sdk~=0.34.1",
    "pandas>=1.3.0"
]

[project.scripts]
tap-excel = "tap_excel:main"
